+++ packages/zocker/src/lib/v4/generators/index.ts
@@ -22,7 +22,7 @@ export { ReadonlyGenerator } from "./readonly.js";
 export { StringGenerator } from "./string/plain.js";
 export { CUIDGenerator, CUID2Generator } from "./string/cuid.js";
 export { IPv4Generator, IPv6Generator } from "./string/ip.js";
-export { UUIDGenerator } from "./string/uuid.js";
+export { UUIDGenerator, GUIDGenerator } from "./string/uuid.js";
 export { E164Generator } from "./string/e164.js";
 export { EmailGenerator } from "./string/email.js";
 export {
@@ -38,3 +38,5 @@ export { URLGenerator } from "./string/url.js";
 export { PipeGenerator } from "./pipe.js";
 export { EmojiGenerator } from "./string/emoji.js";
 export { Base64Generator, Base64URLGenerator } from "./string/base64.js";
+export { KSUIDGenerator } from "./string/ksuid.js";
+export { XIDGenerator } from "./string/xid.js";