+++ packages/zocker/CHANGELOG.md
@@ -1,5 +1,21 @@
 # zocker
 
+## 2.0.3
+
+### Patch Changes
+
+- 7a97a5a: Change `z.date` generation so that more reasonable values are generated
+- 08d099a: Implement Generator for the undocumented `z.xid()` schema
+
+## 2.0.2
+
+### Patch Changes
+
+- c3c2496: Implemet `z.ksuid()` generator
+- f69fd6d: Fix incorrect format when generating values for `z.iso.datetime({ offset: true })`
+- 0101584: fix: Legacy string formats, such as `z.string().email()` are supported again. However, you should still upgrade to the dedicated schemas, such as `z.email()` as soon as possible.
+- 2994db4: Implement `z.guid` generator
+
 ## 2.0.1
 
 ### Patch Changes